source-sqlserver: Move 'cdcCreateCaptureInstance' function
This belongs over in `replication.go` since we will actually be
calling it during normal replication operation really soon.
