+++ source-sqlserver/replication.go
@@ -8,6 +8,7 @@ import (
 	"encoding/json"
 	"errors"
 	"fmt"
+	"hash/crc32"
 	"strings"
 	"sync"
 	"time"
@@ -605,6 +606,50 @@ func cdcCleanupChangeTable(ctx context.Context, conn *sql.DB, instanceName strin
 		return fmt.Errorf("error cleaning capture instance %q below LSN %X: %w", instanceName, belowLSN, err)
 	}
 	return nil
+}
+
+func cdcCreateCaptureInstance(ctx context.Context, conn *sql.DB, schema, table, username string) (string, error) {
+	// SQL Server table names may be up to 128 characters, but capture instance names must
+	// be at most 100 characters and we have other information to cram in there. The names
+	// must be unique, but other than that they might as well be random strings, the logic
+	// around capture instance selection is based on the (schema, table, instance) info in
+	// `cdc.change_tables` and doesn't care about the names.
+	//
+	// So this logic is just trying to give them a reasonably unique name which also tells
+	// us some useful facts, like as much of the name as we can accomodate and when it was
+	// created. The timestamp also helps ensure that successive capture instances from the
+	// same table are always unique.
+
+	// Human-readable prefix of the name, truncated to 64 characters
+	var prefix = schema + "_" + table
+	if len(prefix) > 64 {
+		prefix = prefix[:64]
+	}
+
+	// A 32-bit hash of the full name provides more uniqueness if the truncated prefixes match for multiple tables.
+	var hash = crc32.ChecksumIEEE([]byte(fmt.Sprintf("%s_%s", schema, table)))
+
+	// The full instance name is the 64 bytes prefix, 8 bytes of hash in hex, 10 bytes of Unix timestamp
+	// as a decimal number, and two underscores for a total of 84 bytes, well under the limit of 100.
+	var instanceName = fmt.Sprintf("%s_%08X_%d", prefix, hash, time.Now().Unix())
+	log.WithFields(log.Fields{
+		"schema":   schema,
+		"table":    table,
+		"instance": instanceName,
+	}).Debug("creating new capture instance")
+
+	// According to https://learn.microsoft.com/en-us/sql/relational-databases/system-stored-procedures/sys-sp-cdc-enable-table-transact-sql
+	// the `sys.sp_cdc_enable_table` procedure "Requires membership in the db_owner fixed database role."
+	//
+	// Thus unless we're running as the superuser or the 'flow_capture' user has been explicitly granted
+	// this role we should expect this to fail. We always try to create capture instances during validation
+	// because it hurts nothing if we're rejected, and if the relevant table-alteration handling option is
+	// enabled we verify as a prerequisite that we have 'db_owner' so this shouldn't fail there.
+	const query = `EXEC sys.sp_cdc_enable_table @source_schema = @p1, @source_name = @p2, @role_name = @p3, @capture_instance = @p4;`
+	if _, err := conn.ExecContext(ctx, query, schema, table, username, instanceName); err != nil {
+		return "", fmt.Errorf("error creating capture instance %q: %w", instanceName, err)
+	}
+	return instanceName, nil
 }
 
 func (db *sqlserverDatabase) ReplicationDiagnostics(ctx context.Context) error {
